Move UnauthenticatedException to security spi near to other security exceptions
