+++ cdap-app-fabric/src/main/java/io/cdap/cdap/app/guice/UnsupportedExploreClient.java
@@ -20,7 +20,6 @@ import com.google.common.util.concurrent.ListenableFuture;
 import io.cdap.cdap.api.dataset.DatasetSpecification;
 import io.cdap.cdap.api.dataset.lib.PartitionKey;
 import io.cdap.cdap.common.ServiceUnavailableException;
-import io.cdap.cdap.common.UnauthenticatedException;
 import io.cdap.cdap.explore.client.ExploreClient;
 import io.cdap.cdap.explore.client.ExploreExecutionResult;
 import io.cdap.cdap.explore.service.ExploreException;
@@ -28,6 +27,7 @@ import io.cdap.cdap.explore.service.MetaDataInfo;
 import io.cdap.cdap.proto.NamespaceMeta;
 import io.cdap.cdap.proto.id.DatasetId;
 import io.cdap.cdap.proto.id.NamespaceId;
+import io.cdap.cdap.security.spi.authentication.UnauthenticatedException;
 
 import java.io.IOException;
 import java.util.List;
