+++ cdap-cli/src/main/java/io/cdap/cdap/cli/completer/element/AppIdCompleter.java
@@ -20,8 +20,8 @@ import com.google.common.base.Supplier;
 import io.cdap.cdap.cli.CLIConfig;
 import io.cdap.cdap.cli.completer.StringsCompleter;
 import io.cdap.cdap.client.ApplicationClient;
-import io.cdap.cdap.common.UnauthenticatedException;
 import io.cdap.cdap.proto.ApplicationRecord;
+import io.cdap.cdap.security.spi.authentication.UnauthenticatedException;
 import io.cdap.cdap.security.spi.authorization.UnauthorizedException;
 
 import java.io.IOException;
