+++ cdap-client-tests/src/test/java/io/cdap/cdap/client/MetaClientAvailableTest.java
@@ -18,7 +18,7 @@ package io.cdap.cdap.client;
 
 import io.cdap.cdap.client.config.ClientConfig;
 import io.cdap.cdap.client.config.ConnectionConfig;
-import io.cdap.cdap.common.UnauthenticatedException;
+import io.cdap.cdap.security.spi.authentication.UnauthenticatedException;
 import io.cdap.cdap.security.spi.authorization.UnauthorizedException;
 import io.cdap.http.AbstractHttpHandler;
 import io.cdap.http.HttpResponder;
