+++ cdap-client-tests/src/test/java/io/cdap/cdap/client/MetaClientTestRun.java
@@ -19,10 +19,10 @@ package io.cdap.cdap.client;
 import com.google.common.base.Charsets;
 import com.google.common.io.Resources;
 import io.cdap.cdap.client.common.ClientTestBase;
-import io.cdap.cdap.common.UnauthenticatedException;
 import io.cdap.cdap.common.conf.Constants;
 import io.cdap.cdap.proto.ConfigEntry;
 import io.cdap.cdap.proto.Version;
+import io.cdap.cdap.security.spi.authentication.UnauthenticatedException;
 import io.cdap.cdap.security.spi.authorization.UnauthorizedException;
 import io.cdap.cdap.test.XSlowTests;
 import org.junit.Assert;
