+++ cdap-client-tests/src/test/java/io/cdap/cdap/client/NamespaceClientTestRun.java
@@ -20,10 +20,10 @@ import io.cdap.cdap.client.common.ClientTestBase;
 import io.cdap.cdap.common.AlreadyExistsException;
 import io.cdap.cdap.common.BadRequestException;
 import io.cdap.cdap.common.NotFoundException;
-import io.cdap.cdap.common.UnauthenticatedException;
 import io.cdap.cdap.common.id.Id;
 import io.cdap.cdap.proto.NamespaceMeta;
 import io.cdap.cdap.proto.id.NamespaceId;
+import io.cdap.cdap.security.spi.authentication.UnauthenticatedException;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
