+++ cdap-client/src/main/java/io/cdap/cdap/client/ArtifactClient.java
@@ -34,7 +34,6 @@ import io.cdap.cdap.common.ArtifactNotFoundException;
 import io.cdap.cdap.common.ArtifactRangeNotFoundException;
 import io.cdap.cdap.common.BadRequestException;
 import io.cdap.cdap.common.NotFoundException;
-import io.cdap.cdap.common.UnauthenticatedException;
 import io.cdap.cdap.internal.io.SchemaTypeAdapter;
 import io.cdap.cdap.proto.artifact.ApplicationClassInfo;
 import io.cdap.cdap.proto.artifact.ApplicationClassSummary;
@@ -42,6 +41,7 @@ import io.cdap.cdap.proto.artifact.PluginInfo;
 import io.cdap.cdap.proto.artifact.PluginSummary;
 import io.cdap.cdap.proto.id.ArtifactId;
 import io.cdap.cdap.proto.id.NamespaceId;
+import io.cdap.cdap.security.spi.authentication.UnauthenticatedException;
 import io.cdap.cdap.security.spi.authorization.UnauthorizedException;
 import io.cdap.common.ContentProvider;
 import io.cdap.common.http.HttpMethod;
