+++ cdap-client/src/main/java/io/cdap/cdap/client/AuthorizationClient.java
@@ -24,7 +24,6 @@ import io.cdap.cdap.client.config.ClientConfig;
 import io.cdap.cdap.client.util.RESTClient;
 import io.cdap.cdap.common.FeatureDisabledException;
 import io.cdap.cdap.common.NotFoundException;
-import io.cdap.cdap.common.UnauthenticatedException;
 import io.cdap.cdap.common.conf.Constants;
 import io.cdap.cdap.proto.codec.EntityIdTypeAdapter;
 import io.cdap.cdap.proto.id.EntityId;
@@ -35,6 +34,7 @@ import io.cdap.cdap.proto.security.Principal;
 import io.cdap.cdap.proto.security.Privilege;
 import io.cdap.cdap.proto.security.RevokeRequest;
 import io.cdap.cdap.proto.security.Role;
+import io.cdap.cdap.security.spi.authentication.UnauthenticatedException;
 import io.cdap.cdap.security.spi.authorization.AbstractAuthorizer;
 import io.cdap.cdap.security.spi.authorization.AlreadyExistsException;
 import io.cdap.cdap.security.spi.authorization.UnauthorizedException;
