+++ cdap-client/src/main/java/io/cdap/cdap/client/MetaClient.java
@@ -21,9 +21,9 @@ import com.google.common.reflect.TypeToken;
 import io.cdap.cdap.api.annotation.Beta;
 import io.cdap.cdap.client.config.ClientConfig;
 import io.cdap.cdap.client.util.RESTClient;
-import io.cdap.cdap.common.UnauthenticatedException;
 import io.cdap.cdap.proto.ConfigEntry;
 import io.cdap.cdap.proto.Version;
+import io.cdap.cdap.security.spi.authentication.UnauthenticatedException;
 import io.cdap.cdap.security.spi.authorization.UnauthorizedException;
 import io.cdap.common.http.HttpMethod;
 import io.cdap.common.http.HttpResponse;
