+++ cdap-client/src/main/java/io/cdap/cdap/client/MetricsClient.java
@@ -28,12 +28,12 @@ import io.cdap.cdap.api.annotation.Beta;
 import io.cdap.cdap.api.metrics.RuntimeMetrics;
 import io.cdap.cdap.client.config.ClientConfig;
 import io.cdap.cdap.client.util.RESTClient;
-import io.cdap.cdap.common.UnauthenticatedException;
 import io.cdap.cdap.common.conf.Constants;
 import io.cdap.cdap.common.metrics.MetricsTags;
 import io.cdap.cdap.proto.MetricQueryResult;
 import io.cdap.cdap.proto.MetricTagValue;
 import io.cdap.cdap.proto.id.ServiceId;
+import io.cdap.cdap.security.spi.authentication.UnauthenticatedException;
 import io.cdap.cdap.security.spi.authorization.UnauthorizedException;
 import io.cdap.common.http.HttpRequest;
 import io.cdap.common.http.HttpResponse;
