+++ cdap-client/src/main/java/io/cdap/cdap/client/MonitorClient.java
@@ -24,11 +24,11 @@ import io.cdap.cdap.client.util.RESTClient;
 import io.cdap.cdap.common.BadRequestException;
 import io.cdap.cdap.common.NotFoundException;
 import io.cdap.cdap.common.ServiceNotEnabledException;
-import io.cdap.cdap.common.UnauthenticatedException;
 import io.cdap.cdap.proto.Instances;
 import io.cdap.cdap.proto.SystemServiceLiveInfo;
 import io.cdap.cdap.proto.SystemServiceMeta;
 import io.cdap.cdap.proto.id.SystemServiceId;
+import io.cdap.cdap.security.spi.authentication.UnauthenticatedException;
 import io.cdap.cdap.security.spi.authorization.UnauthorizedException;
 import io.cdap.common.http.HttpMethod;
 import io.cdap.common.http.HttpRequest;
