+++ cdap-client/src/main/java/io/cdap/cdap/client/SecureStoreClient.java
@@ -26,10 +26,10 @@ import io.cdap.cdap.common.AlreadyExistsException;
 import io.cdap.cdap.common.NamespaceNotFoundException;
 import io.cdap.cdap.common.SecureKeyAlreadyExistsException;
 import io.cdap.cdap.common.SecureKeyNotFoundException;
-import io.cdap.cdap.common.UnauthenticatedException;
 import io.cdap.cdap.proto.id.NamespaceId;
 import io.cdap.cdap.proto.id.SecureKeyId;
 import io.cdap.cdap.proto.security.SecureKeyCreateRequest;
+import io.cdap.cdap.security.spi.authentication.UnauthenticatedException;
 import io.cdap.cdap.security.spi.authorization.UnauthorizedException;
 import io.cdap.common.http.HttpMethod;
 import io.cdap.common.http.HttpResponse;
