+++ cdap-client/src/main/java/io/cdap/cdap/client/WorkflowClient.java
@@ -24,7 +24,6 @@ import io.cdap.cdap.api.workflow.WorkflowToken;
 import io.cdap.cdap.client.config.ClientConfig;
 import io.cdap.cdap.client.util.RESTClient;
 import io.cdap.cdap.common.NotFoundException;
-import io.cdap.cdap.common.UnauthenticatedException;
 import io.cdap.cdap.proto.DatasetSpecificationSummary;
 import io.cdap.cdap.proto.WorkflowNodeStateDetail;
 import io.cdap.cdap.proto.WorkflowTokenDetail;
@@ -33,6 +32,7 @@ import io.cdap.cdap.proto.codec.WorkflowTokenDetailCodec;
 import io.cdap.cdap.proto.codec.WorkflowTokenNodeDetailCodec;
 import io.cdap.cdap.proto.id.NamespaceId;
 import io.cdap.cdap.proto.id.ProgramRunId;
+import io.cdap.cdap.security.spi.authentication.UnauthenticatedException;
 import io.cdap.cdap.security.spi.authorization.UnauthorizedException;
 import io.cdap.common.http.HttpMethod;
 import io.cdap.common.http.HttpResponse;
