+++ cdap-client/src/main/java/io/cdap/cdap/client/util/RESTClient.java
@@ -21,8 +21,8 @@ import com.google.common.collect.Lists;
 import com.google.inject.Inject;
 import io.cdap.cdap.client.config.ClientConfig;
 import io.cdap.cdap.client.exception.DisconnectedException;
-import io.cdap.cdap.common.UnauthenticatedException;
 import io.cdap.cdap.security.authentication.client.AccessToken;
+import io.cdap.cdap.security.spi.authentication.UnauthenticatedException;
 import io.cdap.cdap.security.spi.authorization.UnauthorizedException;
 import io.cdap.common.http.HttpMethod;
 import io.cdap.common.http.HttpRequest;
