+++ cdap-explore-client/src/main/java/io/cdap/cdap/explore/client/AbstractExploreClient.java
@@ -32,7 +32,6 @@ import com.google.gson.Gson;
 import io.cdap.cdap.api.dataset.DatasetSpecification;
 import io.cdap.cdap.api.dataset.lib.PartitionKey;
 import io.cdap.cdap.common.ServiceUnavailableException;
-import io.cdap.cdap.common.UnauthenticatedException;
 import io.cdap.cdap.explore.service.Explore;
 import io.cdap.cdap.explore.service.ExploreException;
 import io.cdap.cdap.explore.service.HandleNotFoundException;
@@ -45,6 +44,7 @@ import io.cdap.cdap.proto.QueryStatus;
 import io.cdap.cdap.proto.id.DatasetId;
 import io.cdap.cdap.proto.id.NamespaceId;
 import io.cdap.cdap.security.spi.authentication.SecurityRequestContext;
+import io.cdap.cdap.security.spi.authentication.UnauthenticatedException;
 import org.apache.twill.common.Threads;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
