+++ cdap-explore-client/src/main/java/io/cdap/cdap/explore/client/ExploreClient.java
@@ -20,12 +20,12 @@ import com.google.common.util.concurrent.ListenableFuture;
 import io.cdap.cdap.api.dataset.DatasetSpecification;
 import io.cdap.cdap.api.dataset.lib.PartitionKey;
 import io.cdap.cdap.common.ServiceUnavailableException;
-import io.cdap.cdap.common.UnauthenticatedException;
 import io.cdap.cdap.explore.service.ExploreException;
 import io.cdap.cdap.explore.service.MetaDataInfo;
 import io.cdap.cdap.proto.NamespaceMeta;
 import io.cdap.cdap.proto.id.DatasetId;
 import io.cdap.cdap.proto.id.NamespaceId;
+import io.cdap.cdap.security.spi.authentication.UnauthenticatedException;
 
 import java.io.Closeable;
 import java.util.List;
