+++ cdap-explore-client/src/main/java/io/cdap/cdap/explore/client/ExploreHttpClient.java
@@ -30,7 +30,6 @@ import io.cdap.cdap.api.dataset.ExploreProperties;
 import io.cdap.cdap.api.dataset.lib.PartitionKey;
 import io.cdap.cdap.api.dataset.lib.PartitionedFileSetArguments;
 import io.cdap.cdap.common.ServiceUnavailableException;
-import io.cdap.cdap.common.UnauthenticatedException;
 import io.cdap.cdap.common.conf.Constants;
 import io.cdap.cdap.common.http.DefaultHttpRequestConfig;
 import io.cdap.cdap.explore.service.Explore;
@@ -53,6 +52,7 @@ import io.cdap.cdap.proto.TableInfo;
 import io.cdap.cdap.proto.TableNameInfo;
 import io.cdap.cdap.proto.id.DatasetId;
 import io.cdap.cdap.proto.id.NamespaceId;
+import io.cdap.cdap.security.spi.authentication.UnauthenticatedException;
 import io.cdap.common.http.HttpMethod;
 import io.cdap.common.http.HttpRequest;
 import io.cdap.common.http.HttpRequestConfig;
