+++ cdap-integration-test/src/main/java/io/cdap/cdap/test/remote/RemoteSparkManager.java
@@ -21,11 +21,11 @@ import io.cdap.cdap.client.config.ConnectionConfig;
 import io.cdap.cdap.client.util.RESTClient;
 import io.cdap.cdap.common.NotFoundException;
 import io.cdap.cdap.common.ServiceUnavailableException;
-import io.cdap.cdap.common.UnauthenticatedException;
 import io.cdap.cdap.common.discovery.URIScheme;
 import io.cdap.cdap.common.service.ServiceDiscoverable;
 import io.cdap.cdap.common.utils.Tasks;
 import io.cdap.cdap.proto.id.ProgramId;
+import io.cdap.cdap.security.spi.authentication.UnauthenticatedException;
 import io.cdap.cdap.security.spi.authorization.UnauthorizedException;
 import io.cdap.cdap.test.AbstractProgramManager;
 import io.cdap.cdap.test.ApplicationManager;
