+++ cdap-integration-test/src/test/java/io/cdap/cdap/test/IntegrationTestBaseTest.java
@@ -22,10 +22,10 @@ import io.cdap.cdap.StandaloneTester;
 import io.cdap.cdap.api.dataset.DatasetAdmin;
 import io.cdap.cdap.client.ApplicationClient;
 import io.cdap.cdap.client.config.ClientConfig;
-import io.cdap.cdap.common.UnauthenticatedException;
 import io.cdap.cdap.proto.NamespaceMeta;
 import io.cdap.cdap.proto.ProgramRunStatus;
 import io.cdap.cdap.proto.id.NamespaceId;
+import io.cdap.cdap.security.spi.authentication.UnauthenticatedException;
 import io.cdap.cdap.security.spi.authorization.UnauthorizedException;
 import io.cdap.common.http.HttpMethod;
 import org.junit.Assert;
